Refactor before PR still needs unit testing
