+++ src/test/java/app/util/PartCalculatorTest.java
@@ -183,40 +183,4 @@ class PartCalculatorTest
         Carport carport420x420WithOutShed = new Carport(0, 600, 600, RoofType.FLAT, null);
         assertEquals(1, PartCalculator.calculateNumberOfBracketScrewsNeeded(carport420x420WithOutShed, screwsPerPackage));
     }
-
-    @Test
-    void testCalculateCenterPostPlacementWith220cmShed()
-    {
-        Carport carport = new Carport(0, 780, 600, RoofType.FLAT, new Shed(0, 220, 220, ShedPlacement.FULL_WIDTH));
-        int centerpostPlacement = PartCalculator.calculateCenterPostPlacement(carport);
-
-        assertEquals(410, centerpostPlacement);
-    }
-
-    @Test
-    void testCalculateCenterPostPlacementWith270cmShed()
-    {
-        Carport carport = new Carport(0, 780, 600, RoofType.FLAT, new Shed(0, 270, 220, ShedPlacement.FULL_WIDTH));
-        int centerpostPlacement = PartCalculator.calculateCenterPostPlacement(carport);
-
-        assertEquals(380, centerpostPlacement);
-    }
-
-    @Test
-    void testCalculateCenterPostPlacementWith380cmShed()
-    {
-        Carport carport = new Carport(0, 780, 600, RoofType.FLAT, new Shed(0, 380, 220, ShedPlacement.FULL_WIDTH));
-        int centerpostPlacement = PartCalculator.calculateCenterPostPlacement(carport);
-
-        assertEquals(480, centerpostPlacement);
-    }
-
-    @Test
-    void testCalculateCenterPostPlacementWithShedLengthEqualToCenterPost()
-    {
-        Carport carport = new Carport(0, 780, 600, RoofType.FLAT, new Shed(0, 340, 220, ShedPlacement.FULL_WIDTH));
-        int centerpostPlacement = PartCalculator.calculateCenterPostPlacement(carport);
-
-        assertEquals(410, centerpostPlacement);
-    }
 }
